fix(个人主页): 当页切换路由时入口不会添加
