+++ packages/web-extension/src/content-scripts/utils/getInformation.ts
@@ -1,0 +1,5 @@
+export function getCurrentUserId() {
+    const match = window.location.pathname.match(/\/user\/(\d+)/);
+    return match ? match[1] : ""
+}
+
